Add files via upload
init-graph.py is the current working copy of the code to read in CSV files and display the outputs in a Dash webapp. This code only reads in and graphs test data Test.csv. Future iterations will allow for accumulating data from multiple files and add a feature to select the folder which selects the sensor